Module Emit_alarm

type hook = {

hook : 'a. 'a Operator.Alarm.t -> Tracelog.location list -> unit;

}

val register_alarm_hook : hook -> unit

Register a hook to be called when a new alarm has been emitted. The hook will be called with an alarm and the stack of tracelog locations.

val emit_alarm : 'a Operator.Alarm.t -> unit

Call all the registered hooks, in order, on the emitted alarm.

val reset_alarms : unit -> unit

Reset the alarm logs, this should be used only for post analysis